Revert "Move compareVersions to libutil"
This reverts commit 23347cb3a946cc0adb10ecb8566d28ddd8f56cf9.

we no longer need this with dd9bb11d0d38139bb32411170403171c4c92f8cf
