+++ src/libstore/names.hh
@@ -3,7 +3,6 @@
 #include <memory>
 
 #include "types.hh"
-#include "versions.hh"
 #include <regex>
 
 namespace nix {
@@ -25,6 +24,9 @@ private:
 
 typedef list<DrvName> DrvNames;
 
+string nextComponent(string::const_iterator & p,
+    const string::const_iterator end);
+int compareVersions(const string & v1, const string & v2);
 DrvNames drvNamesFromArgs(const Strings & opArgs);
 
 }
